Fully functional vending machine //TODO: User wallet

+++ VendingMachine/Models/Product.cs
@@ -7,10 +7,9 @@ namespace VendingMachineApp.Models
     class Product
     {
         public string Name { get; set; }
-        public double Price { get; set; }       
+        public decimal Price { get; set; }       
         public int Amount { get; set; }
-
-        public Product(string name, double price, int amount)
+        public Product(string name, decimal price, int amount)
         {
             this.Name = name;
             this.Price = price;
